👌 Updating code due to code review changes
* remove JSONDecoder extension
* replace getQueryItems func with property
