+++ CriticalMass/TwitterRequest.swift
@@ -5,13 +5,12 @@ struct TwitterRequest: APIRequestDefining {
     var endpoint: Endpoint = .twitter
     var headers: HTTPHeaders?
     var httpMethod: HTTPMethod = .get
+    var queryItems: [URLQueryItem]?
 
     func parseResponse(data: Data) throws -> ResponseDataType {
         let decoder = JSONDecoder.decoder(dateDecodingStrategy: .formatted(.twitterDateFormatter))
         return try decoder.decode(ResponseDataType.self, from: data)
     }
-
-    func getQueryItems() -> [URLQueryItem]? { nil }
 }
 
 private extension DateFormatter {
